Relocate button component to shared directory
