--- packages/client/src/components/Button/Button.stories.tsx
@@ -1,22 +1,0 @@
-import React from "react";
-import { Story, Meta } from "@storybook/react";
-import Button, { ButtonProps } from "./Button";
-
-export default {
-  title: "Button",
-  component: Button,
-} as Meta;
-
-const Template: Story<ButtonProps> = (args) => <Button {...args} />;
-
-export const Primary = Template.bind({});
-Primary.args = {
-  color: "primary",
-  children: "Primary Button",
-};
-
-export const Secondary = Template.bind({});
-Secondary.args = {
-  color: "secondary",
-  children: "Secondary Button",
-};
