+++ packages/client/src/shared-components/button/index.ts
@@ -1,0 +1,2 @@
+export * from './button';
+export { default } from './button';
